Proyecto/backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
pymongo==4.6.1
pytest==7.4.3
PyJWT==2.8.0
